# Conflicts:
#	Cartfile.resolved
#	Podfile
#	Podfile.lock
#	cybexMobile/Info.plist
#	cybexMobile/Modules/KLine/Views/CBKLineMainView.swift
#	cybexMobile/UI/YourPortfolioHeadView.swift
#	cybexMobile/ViewControllers/Account/OpenedOrders/View/AccountOpenedOrdersView.swift
#	cybexMobile/ViewControllers/Account/Portfolio/Views/YourPorfolioView.swift
#	cybexMobile/ViewControllers/Entry/Register/RegisterViewController.swift
#	cybexMobile/ViewControllers/Home/HomeViewController.swift
#	cybexMobile/ViewControllers/Home/Market/MarketViewController.swift
#	cybexMobile/ViewControllers/Home/Market/Views/PairCardView.swift
#	cybexMobile/ViewControllers/Home/Market/Views/PairDetailView.swift
#	cybexMobile/ViewControllers/Home/Market/Views/PairListHorizantalView.swift
#	cybexMobile/ViewControllers/Home/Views/HomeContentView.swift
#	cybexMobile/ViewControllers/Trade/Business/views/BusinessTitleItemView.swift
#	cybexMobile/ViewControllers/Trade/Business/views/BusinessTitleView.swift
#	cybexMobile/ViewControllers/Trade/Views/MyOpenedOrdersView.swift
#	cybexMobile/ViewControllers/Trade/Views/TradeHistoryView.swift
#	cybexMobile/ViewControllers/Trade/Views/TradeLineView.swift

+++ cybexMobile/ViewControllers/Account/OpenedOrders/OpenedOrdersCoordinator.swift
@@ -8,6 +8,7 @@
 
 import UIKit
 import ReSwift
+import cybex_ios_core_cpp
 
 // 跳转
 protocol OpenedOrdersCoordinatorProtocol {
@@ -57,7 +58,7 @@ extension OpenedOrdersCoordinator: OpenedOrdersStateManagerProtocol {
       if success {
         blockchainParams { (blockchain_params) in
           guard let asset = app_data.assetInfo[assetID] else {return}
-            if let jsonStr = BitShareCoordinator.cancelLimitOrder(blockchain_params.block_num, block_id: blockchain_params.block_id, expiration: Date().timeIntervalSince1970 + 10 * 3600, chain_id: blockchain_params.chain_id, user_id: userid.getID, order_id: orderID.getID, fee_id: assetID.getID, fee_amount: Int64(Int32(amount.doubleValue * pow(10, asset.precision.double)))) {
+            if let jsonStr = BitShareCoordinator.cancelLimitOrder(blockchain_params.block_num, block_id: blockchain_params.block_id, expiration: Date().timeIntervalSince1970 + 10 * 3600, chain_id: blockchain_params.chain_id, user_id: userid.getID, order_id: orderID.getID, fee_id: assetID.getID, fee_amount: Int64(amount.doubleValue * pow(10, asset.precision.double))) {
             let request = BroadcastTransactionRequest(response: { (data) in
               if String(describing: data) == "<null>" {
                 callback(true)
